🐛 Fix: 스크롤, 날짜 포맷, 루시드 임포트 수정
